split domain/type.ts
