+++ src/common/utils/mainPages.ts
@@ -1,7 +1,7 @@
 import React from "react";
 import { About } from "../../about/components/About";
-import { aboutData, portfolioData, resumeData } from "../../domain/data";
-import { Data } from "../../domain/type";
+import { aboutData, portfolioData, resumeData } from "../../data";
+import { Data } from "../../domain";
 import { Portfolio } from "../../portfolio/components/Portfolio";
 import { Resume } from "../../resume/components/Resume";
