+++ src/domain/header.ts
@@ -1,0 +1,1 @@
+export type HeaderDataType = { name: string; description: string };
