+++ src/domain/index.ts
@@ -1,0 +1,10 @@
+import { AboutData } from "./about";
+import { PortfolioData } from "./portfolio";
+import { ResumeData } from "./resume";
+
+export type Data = AboutData | ResumeData | PortfolioData;
+
+export * from "./header";
+export * from "./about";
+export * from "./resume";
+export * from "./portfolio";
